Refactor: Move more functionality outside of `etheno.py`
