+++ etheno/ganache.py
@@ -2,30 +2,12 @@
 
 import atexit
 import json
-import socket
 import subprocess
 import time
 from urllib.request import urlopen
-from urllib.error import HTTPError, URLError
 
-def webserver_is_up(url):
-    try:
-        return urlopen(url).getcode()
-    except HTTPError:
-        # This means we connected
-        return True
-    except URLError:
-        return False
-
-def is_port_free(port):
-    sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-    return sock.connect_ex(('127.0.0.1', port)) != 0
-
-def find_open_port(starting_port=1025):
-    for port in range(starting_port, 65536):
-        if is_port_free(port):
-            return port
-    return -1
+from .client import SelfPostingClient
+from .utils import is_port_free
 
 class RpcHttpProxy(object):
     def __init__(self, urlstring):
@@ -83,3 +65,10 @@ if __name__ == "__main__":
         'params': [],
         'id': 1
     }))
+
+class GanacheClient(SelfPostingClient):
+    def wait_until_running(self):
+        while is_port_free(self.client.port):
+            time.sleep(0.25)
+    def shutdown(self):
+        self.client.stop()
